Implemented basic breadcrumb capability
Added breadcrumbs to existing views. Breadcrumbs are automatic except for on the detail pages when it needs to be added manually. This is so that we can leverage the data loaded from the server without calling it twice (we may be able to automate this and improve it).

Fixed the breadcrumb container to the site header.

Added highlighting to active area i.e. Projects